requirements.txt
paddleocr>=2.7.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
pdf2image>=1.16.0
easyocr>=1.7.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
langdetect>=1.0.9
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0